Added functionality for AlphaPose generated json files. Some parameters are also updated.
A number of class methods have been added to Pose2D and Poses2D to allow for data to be read from AlphaPose json files. Genearlization of these methods to allow for the use of multiple model json outputs to be read with little user input would ideal, but for now the functionality is available for both OpenPose and AlphaPose with slight modifications of parameters. Currently, the body part labeling parameters are set for Alpha Pose.